add some test cases for pour point-related functions
