+++ setup.py
@@ -49,6 +49,7 @@ setup(
         'natcap',
         'natcap.invest',
         'natcap.invest.coastal_blue_carbon',
+        'natcap.invest.delineateit',
         'natcap.invest.finfish_aquaculture',
         'natcap.invest.fisheries',
         'natcap.invest.hydropower',
